+++ server/db/db.js
@@ -1287,6 +1287,38 @@ async function getCurrentTime(pgClient) {
   }
 }
 
+// Checks the database connection, if not able to connect will throw an error
+async function getCurrentTimeForHealthCheck() {
+  if (helpers.isDbLogging()) {
+    helpers.log(`STARTED: getCurrentTimeForHealthCheck`)
+  }
+
+  let pgClient = null
+
+  try {
+    pgClient = await pool.connect()
+    if (helpers.isDbLogging()) {
+      helpers.log(`CONNECTED: getCurrentTimeForHealthCheck`)
+    }
+
+    const results = await pgClient.query(`SELECT NOW()`)
+    return results.rows[0].now
+  } catch (err) {
+    helpers.logError(`Error running the getCurrentTimeForHealthCheck query: ${err}`)
+    throw err
+  } finally {
+    try {
+      pgClient.release()
+    } catch (err) {
+      helpers.logError(`getCurrentTimeForHealthCheck: Error releasing client: ${err}`)
+    }
+
+    if (helpers.isDbLogging()) {
+      helpers.log(`COMPLETED: getCurrentTimeForHealthCheck`)
+    }
+  }
+}
+
 async function close() {
   try {
     await pool.end()
@@ -1348,6 +1380,7 @@ module.exports = {
   getButtons,
   getButtonWithSerialNumber,
   getCurrentTime,
+  getCurrentTimeForHealthCheck,
   getDataForExport,
   getHistoricAlertsByAlertApiKey,
   getClients,